broker: add content_cache_backing_loaded()
Problem: In the near future, code in the broker outside of the
content cache will need to determine if a backing module is loaded.
However, there is no way to tell if one is loaded outside of the
content cache.

Solution: Add a new function content_cache_backing_loaded() that
will return true if a backing module is loaded.