flux-job: point users to flux-jobs(1)
Problem: flux job list is too easy to confuse with flux-jobs(1),
but the former is a test tool that emits JSON.

Omit the list, list-inactive, and list-ids subcommands from help output.
If output is a tty, assume a human is trying to ingest JSON (blech!)
and throw a fatal error that instructs them to try flux-jobs(1).

Fixes #4497