update fluid check to check explicitly for utf-8
Some multibyte locales do not support the florin symbol, and we were
checking for width rather than encoding.  This check looks explicitly
for UTF-8 support in the current locale, and because it turns out that
python interprets C as C.UTF-8, and propagates that through to other
code it calls.  To be extra conservative this brings back the original
check so that we will only print the florin character if we are in a
natively UTF-8 locale with a multibyte character set.

fixes #4504