libflux: plugin: add flux_plugin_get_path()
Problem: The flux_plugin_t class does not have an accessor to get the
currently loaded plugin path. This is not convenient when a plugin
is loaded from a colon-separated path, as the caller can't determine
what actual plugin was loaded.

Add a flux_plugin_get_path(3) accessor to return the path for a plugin
loaded by dlopen(3).