job-manager: support jobtap-query in dependency-after plugin
Problem: There is no easy way to collect the list of outstanding job
dependencies held in the dependency-after jobtap plugin.

Since iterating all active jobs is infeasible and would be inefficient
when only a subset of jobs have dependencies, add a global list of
references which can be used to iterate all active dependencies.

Add a 'plugin.query' handler to the dependency-after plugin which
then places an array of active dependencies, including the jobid,
the antecedent jobid, the type of dependency and the description
(used in the dependency-add event).

This may be useful for tools (e.g. a tool could use this data to
create a visualization of extant dependencies), but also serves as
an example of utilizing the jobtap-query support for plugins.