job-ingest: add job class
Problem: it would facilitate refactoring if 'struct job' could be
moved out of the job-ingest mainline and shared with other source
modules.

Create a job class, and a constructor called job_create_from_request()
that, as a side benefit, allows a large chunk of the submit request
handler to be moved out of job-ingest.

Add a unit test.