job-ingest: add pipeline class
Problem: the validation workcrew will soon be turned into a pipeline
of work crews, but the code is intermixed with the job-info mainline
making it harder to see what is going on.

Create a "pipeline" class that encapsulates the validation work crew
for now, and will be more easily extended to a pipeline without making
the code awful to read.