cmd/flux-jobs: include job state in status output
Problem: Currently the job status outputs "PENDING"
when the job is in the DEPEND or SCHED state and "RUNNING" when the
job is in the RUN or CLEANUP states.  This abstraction of job states
in the job status doesn't give enough detail about what is going
on with a job.

Solution: Update the job status to output the job state when a job
is not inactive.  Output the job result when it is inactive.  In
other words, the list of outputs for the job status is now:

DEPEND, SCHED, RUN, CLEANUP, COMPLETED, FAILED, CANCELED, or TIMEOUT

Update tests for new output accordingly.

Update documentation accordingly.

Fixes #4495