flux-jobs: add --since and --name options
Problem: The flux.job.JobList() interface supports 'since' and 'name'
arguments to filter jobs that went inactive _since_ a given timestamp
or that have a specific _name_, but these features are not available in
flux-jobs.

Add --since=TIME and --name=NAME options to flux-jobs to allow access
to these new job-list features.