flux-jobs: make args.filter a set
Problem: The destination of the --filter option in flux-jobs is a
comma separated string, which makes determining which, if any, filter
options have been specified, and makes it hard to add new options to
the existing value without duplicates.

Change the args.filter object to a Python set instead of a string.
Add a new FilterActionSetUpdate class to handle updating members
of this set from the command line. The option may now be specified
multiple times on the command line.

In places where the set is modified, use set operations instead of
overwriting the value.