broker: fix double decref in checkpoint error path
Problem: In several functions in the checkpoint code, a double
decref of a message reference can occur.  This occurs due to
a direct call to flux_msg_decref() and another which can
occur when the future is destroyed.

Solution: Correct the error and make cleanup easier by placing
forwarding requests into their own functions.  In addition,
set an errstr with an explanation of the error that can be
returned to callers.

of a msg reference pointer could occur in an error path.