pipeline: improve validator bypass
Problem: validator cannot be disabled on the fly and config callback
is rather complex.

Simplify the config callback by creating the validator workcrew
unconditionally, and allowing it to be bypassed independently of that.
This should be easier to work on than the existing logic, and
the overhead of creating the work crew and then bypassing it should
be negligible.

Log one message rather than two upon (re-)configuration.  Update
tests in t2111-job-ingest-configure.t that depended on the old log
message format.