job-ingest: add frobnicator to pipeline
Problem: job-ingest needs to allow jobspec to be modified
before the validator (optionally) checks jobspec feasibility.

Add an ingest pipeline stage that calls flux-job-frobnicator,
a command line utility similar to flux-job-validator.  The pipeline
stage fulfills a future with modified jobspec, which replaces
job->jobspec before it is sent to the validator stage.