libjob: add unwrap_string()
Problem: A function to "unwrap" a string signed by flux-security (or
the flux-core sign_wrap_none()), which works whether flux-core is built
with flux-security or not and always returns a NUL terminated string,
is non-trivial and will be used in at least 2 places. (flux-job and
the job-shell). It would be best if a shared function which can be
separately unit tested was provided, rather than having each user
duplicate the necessary code.

Add unwrap_string() to libjob (for internal use only). The function
is added to libjob since this was the location of the (also internal
use) sign_wrap_none() and sign_unwrap_none().

unwrap_string() will unwrap signed data as a string, always ensuring
there is NUL termination, even if the signature doesn't include it.
The function works whether or not flux-core is built against
flux-security and can optionally verify the signing mechanism is
allowed and that the signing userid matches the current uid.