flux-job: add --original option to flux-job info
Problem: There is no command line utility to fetch the signed J
for a job and return the original, unmodified jobspec that a user
submitted. This may be necessary if the redacted jobspec is one day
stored in the KVS as "jobspec".

Add a new option, --original, to flux-job info which can be used to
translate the "jobspec" key into a fetch and decode of "J". That is

 flux job info --original jobspec

will return the original jobspec submitted by the user by fetching
J and "unwrapping" it by making use of the new unwrap_string()
function.