testsuite: fetch original jobspec in dummy job shell
Problem: The dummy job shell fetches the jobspec via `flux job info
jobspec`, but this won't work if the jobspec stored in the "jobspec"
key is redacted or modified.

Supply the --original flag to `flux job info` to fetch J instead and
decode the result, so the full, originally submitted jobspec is
obtained.