testsuite: enhance dummy job shell
Problem: The dummy job shell logs messages to stdout and doesn't post
a shell.start event. These missing features, especially the second
make it so the dummy shell won't work for some tests.

Add a log() function to print debug messages to stderr instead
of stdout.

Manually post a shell.start event to the exec eventlog so that tests
looking for this event can succeed.