testsuite: signal process group in dummy IMP
Problem: The dummy IMP in the testsuite only signals the job shell
instead of the whole process group. This can cause some test jobs
to hang if the dummy shell is used and it has children in its process
group.

Signal the entire job process group from the dummy IMP kill handler
to avoid this problem.