testsuite: prepare multiuser test for job shell checking J
Problem: The t2404-job-exec-multiuser.t test submits jobs using a
fake userid. Once the job shell begins validating J though, these
tests will fail early in the job shell intiialization. This is fine
for some tests (which are testing that the IMP is invoked), but for
the signal test some kind of job must run.

Switch the 'imp kill' test to use the dummy job shell, which will
not validate J and thus will run the specified tasks.

Modify some other tests to use 'test_might_fail' since jobs that do
not fail now may do so in the future.