testsuite: ensure shell uses and validates J
Problem: The testsuite has no tests that ensure the job shell fetches
J instead of jobspec, and validate it for authenticity and integrity
before continuing with the job.

Add a couple tests to t2602-job-shell.t that modify jobspec and J
and ensure the job shell ignores it in the first case, or fails in
the second.