testsuite: avoid prereq errors in t2404-job-exec-multiuser.t
Problem: The jq prereq test, which checks that HAVE_JQ is set on
tests that use jq, fails when running the dummy job shell in the
multiuser exec tests since the correct per-test sharness environment
variables aren't available from the dummy shell.

Just skip all the tests in t2404-job-exec-multiuser.t when jq is not
available. Since this is done before test_under_flux() is invoked,
the correct environment variables that validate the jq prereq are
available in the dummy shell environment.