broker: log content store errors to LOG_CRIT
Problem: In the event of content store errors, such as ENOSPC
if the disk is full, error messages are logged to LOG_ERR.  This
does not properly signify the severity of the error, as something
like ENOSPC would mean no data is being backed up to disk.

Solution: Log to LOG_CRIT, to signify the importance of this
issue and the need for it to be resolved.