python: add JobFrobnicator class and flux job-frobnicator cmd
Problem: There is a need for a command and Python class similar
to what exists for job validation -- implemented as a command which
can filter and modify jobspec, driven as a worker by the job-ingest
module.

Add a JobFrobnicator class, very similar to the JobValidator class,
which loads a set of plugins. However, instead of running all plugins
in parallel, jobspec is passed through a stack of plugins which can
modify jobspec in place.

A flux-job-frobnicator.py command is provided which, similar to
flux-job-validator.py, can act as a job-ingest worker. Given
jobspec, it will return the modified jobspec to the caller using the
job-ingest worker protocol by default.

Includes a `defaults` plugin which will set defaults in jobspec
based on defaults set by configuration policy.