kvs: support gc-threshold config
Problem: KVS garbage collection is only done when an
administrator runs flux-shutdown and chooses to
garbage collect via the --dump or --gc options.

Solution: Support a kvs gc-threshold configuration option.
This configuration will take an integer count of KVS changes
(the KVS version number or sequence number).  Once the threshold
has been crossed, flux-shutdown will ask the user if they wish to
garbage collect.  Additional options are added to specify yes/no
if the user is scripting with flux-shutdown.

This offers an easy way for administrators to be reminded of garbage
collection on a regular basis.

Fixes #4311