libkvs: pack with JSON_COMPACT
Problem: after using flux_kvs_txn_pack() instead of json_dumps()
to encode jobspec in the KVS, a job-ingest test is failing because
of white space in the jobspec.

Add the JSON_COMPACT flag to json_dumps() in libkvs.  This fixes the
test and more generally avoids transporting and storing unnecessary
characters.