shell: fix hwloc_distrib() usage in affinity plugin
Problem: The job shell uses hwloc_distrib() incorrectly to distribute
cores to tasks when cpu-affinity=per-task is specified, resulting in
strange, uneven distribution of cores to tasks.

Using the source for hwloc-distrib(1) as a reference, distribute cores
to tasks by creating a separate hwloc_obj_t "root" per core in the
current topology, then calling hwloc_distrib() explicitly with those
roots. This seems to force the even distribution of cores to tasks
that is expected.

Fixes #4525