librlist: add RHWLOC_NO_RESTRICT flag
Problem: The librlist rhwloc_local_topoolgy_load() and
rhwloc_local_topology_xml() functions always restrict the generated
topology to the current process cpu binding mask, but this may not
always be desirable.

Add a flag argument to these functions and a RHWLOC_NO_RESTRICT flag
to disable the topology restriction.