resource: add norestrict configuration option
Problem: The hwloc topology XML generated by the resource module is
always restricted to the cpu affinity mask of the broker. This might
not always be desirable, though, for example when the system instance
broker is constrained to a subset of cores by site policy.

Add a 'norestrict' configuration option in the [resource] table to
disable the restricting of XML to the current process affinity mask.
This can be used in a system instance to ensure the entire machine
XML is available to jobs and subinstances.