job-list: return nnodes if jobspec specifies nodes
Problem: job-list does not return nnodes until a job is running.
When a job specifically requests a number of nodes, it may be
useful for job-list to return nnodes so it can be listed in
job listing tools such as flux-jobs.

Solution: If the jobspec specifically lists nodes as a requested
resource, have job-list return nnodes even before it is running.

Fixes #4530