shell: inherit FLUX_F58_FORCE_ASCII from job
Problem: If a user sets FLUX_F58_FORCE_ASCII for a job, then job
artifacts created by the job shell, such as environment variables,
directories, and output files, should obey this variable and use ascii
encoded F58.  However, this is currently not the case since the job
shell is using the environment inherited from the broker and/or IMP,
not the job environment.

Set up the job shell to inherit a subset of environment variables
from the job environment early during startup. At first, add just
FLUX_F58_FORCE_ASCII to this list.