broker/content-mmap: map files into content cache
Problem: the content cache has attractive scalability properties
but since all content lands in the backing store, it cannot be used to
scalably broadcast files without wasting back-end storage.

Add content.mmap-* RPCs that allow files to be mapped with mmap(2) into
the rank 0 broker memory space.  An array of blobrefs is computed over
the file's contents and made available for caching on demand.  Mapped
files can have optional tags.  Tags refer to groups of files for listing
and unmapping.  A file listing contains sufficient information to allow
the files to be reconstructed through the cache.  The listing may be
requested as an array of 'fileref' objects, or an array of blobrefs
that refer to fileref objects.

The 'ephemeral' bit is set on mapped cache entries so that attempts to
store unrelated blobs with the same hash still make it to the backing store
as they did before.

Files can only be mapped on the rank 0 broker.  They remain mapped until
they are explicitly unmapped by request.

Add a cache that maps blobrefs (or rather their raw hash digest) to
mapped regions.  The cache is populated at the time the region is
mapped.  This allows lookups to work at the speed of a hash lookup,
and importantly, it can quickly be determined if a blobref is not
part of an mmapped region so lookup can proceed elsewhere.

Note that blobrefs may be duplicated within a region and across
regions, but duplicate cache entries are not allowed.  If a blob is
already cached, the existing entry remains the definitive source for
the blob.  Since unmapping a region potentially removes cache entries
for valid blobrefs, we must scan all mapped regions and plug any holes
at unmap time.  Unmapping is thus inefficient, however, since only a bulk
unmap (by tag) RPC is supported, this scan only needs to take place after
the bulk operation, not as each region is unmapped.