broker: allow files to mapped into content cache
Problem: the content cache has attractive scalability properties
but it cannot be used to broadcast files without storing them in
the backing store.

Add RPCs that allow a file to be mmapped and an array of blobrefs computed
on its contents.  The rank 0 content cache checks any mmapped regions
before attemptign to faulting a blob in from the backing store.
The cache resulting entry takes a reference on the mmapped region, and
decrements the region's reference count when the cache entry is removed.

The ephemeral bit is set on mapped cache entries so that attempts to
store blobs to the backing store that happen to have the same hash value
as a mapped blob are not subverted.

Currently a region remains mapped until it is unmapped by path, using
another RPC.