content-cache: avoid extra blob copy
Problem: when a cache entry is filled by a load response or a store
request, the message payload is unnecessarily copied to a new buffer
in the cache entry.

Just take a reference on the message and point entry->data directly
to the message payload.  Although there is a small amount of storage
overhead to retain the message envelope, the KVS metadata design packs
small entries into directories so it is expected that most blobs are
large enough that this trade-off is a good one.

In addition this change prepares the cache entry to hold data from
other sources such as mmap(2) without copying.