broker/content-cache: add ephemeral flag
Problem:  mmaped cached entries may hash to the same entry has other
data that needs to be stored in the backing store.  However, mmapped
cached blobs do not need to be stored in the backing store.   There is
no way to differentiate between these two circumstances, so entries
could "squash" one another.

Add an 'ephemeral' cache entry flag that tags non-dirty entries as
not on the backing store.  When a blob with the same hash is stored,
the ephemeral entry is replaced so that it can reach the backing store.