content-cache: add ephemeral flag to cache entries
Problem: storing a blob that happens to have the same hash as a
cached blob that originated from an mmapped file might be "squashed",
thus preventing the blob being stored from reaching the backing
store.

Add an 'ephemeral' cache entry flag that tags non-dirty entries as
not on the backing store.  When a blob with the same hash is stored,
the bit may be cleared and the dirty bit set so that the cache entry
gets written out.