python: add JobspecV1.per_node() constructor
Problem: The existing JobspecV1.from_command() interface does not
easily allow construction of a jobspec specifying a number of nodes
and/or cores along with a number of tasks to launch per node (or core).

Introduce a new JobspecV1.per_node() constructor which is purpose-built
to handle this case. This avoids the need to add more arguments to
from_command() and try to keep backward compatibility.