flux-mini: submit,run: add per-resource allocation options
Problem: A common use case for simple run and submit commands is
to allocate a set of nodes and run a set number of tasks per node,
but this is not easily possible with flux-mini run and submit.

Add a set of "per resource" allocation options to flux-mini run,submit.
If any of these options are used, then the command uses the new
JobspecV1.per_node() constructor to build the submitted jobspec.
This lets users easily decide if they want to run a set number of
tasks with a given amount of resources per task, or allocate a set
of resources with a given number of tasks per resource.