doc: update description of --tasks-per-core in flux-mini(1)
Problem: A lack of exposition on the --tasks-per-core option of
flux-mini may lead users astray in thinking this option is more useful
than it really is.

Move the --tasks-per-core option to the end of the Per-resource
options list, thereby showing it as the lesser sibling of this option
group. Expand the option description to include the many caveats,
warnings, and a final note that most likely this is not the option
you want.