broker: log flux-core version on a client mismatch
Problem: when a client's attempt to connect is rejected due
to a flux-core version mismatch, it would be helpful to log
the actual versions in the server log.

The versions were returned to the client, which logs them, but
not the server.  Restructure the code so that the client and
server log this information.

Use flux_error_t and errprintf() instead of the one-off buffer.

This need was originally noted in #4464.