python: uri: use path to current flux executable in lsf resovler
Problem: The lsf URI resolver plugin executes `ssh flux uri resolve..`
on a compute host during URI resolution, but this may not work if the
remote host has a different version of Flux installed as the system
instance.

Assume a shared filesystem by default and replace `flux` with the path
to the local flux executable by using shutil.which("flux"). As with the
ssh connector, allow this to be overridden with `FLUX_SSH_RCMD`.

Fixes #4558