broker: validate RFC 33 config
Problem: the [queues] and [policy] configuration defined in RFC 33
is not "owned" by any one plugin or flux component, and the frobnicator
(one consumer) is not in a good position to raise a fatal error if it
is incorrect.

Validate the RFC 33 configuration in the broker so it can be checked
before it is distributed to other parts of Flux.  In this position,
the instance will fail to start with a bad config, and flux config
reload will fail on a bad config.  Consumers of these config tables
thus need not exhaustively validate them.

Update t2222-job-manager-limit-job-size.t to expect immediate config
failures instead of failures when the jobtap plugin is loaded.