testsuite: prep for flux-mini --time-limit change
Problem: The default for a unit-less duration specified to the flux-mini
--time-limit=N option will soon change from seconds to minutes, but may
tests in the testsuite assume seconds.

Go through the testsuite and apply "s" units to all bare --time-limit
arguments to ensure they are always interpreted in seconds.

Also remove one test that ensures `-t5` works since after the change
this will result in a 5 minute duration not 5 seconds.