flux-mini: change default units of --time-limit to minutes
Problem: The -t, --time-limit=FSD option of flux-mini is in FSD,
which defaults to seconds when no units are explicitly supplied,
but traditional resource managers with similar options default to
minutes and this has been confusing first-time users.

Change the interpretation of only the --time-limit option to default
to minutes when no units are supplied.

Fixes #4552