librlist: add rlist_from_config()
Problem: There is no user friendly configuration format that can be
used to generate Rv1 for resource configuration.

Add rlist_from_config() to librlist, which can read a simple format
TOML array from `resource.config` where each entry configures an idset
of cores, gpus, and/or properties for a set of hosts. E.g.:

 [[resource.config]]
 hosts = "foo[0-10]"
 cores = "0-3"
 gpus =  "0"

 [[resource.config]]
 hosts = "foo0"
 properties = [ "r0" ]

 [[resource.config]]
 hosts = "foo10"
 properties = [ "r10" ]

Hostnames can appear multiple times and each time the host is updated,
not overridden, which allows a natural configuration from generic to
specific. To avoid confusion, the configuration does not incorporate
ranks, instead ranks are meant to be assigned by the resource module
when it reranks the config.