python: vendor tomli TOML parsing library
Problem: There is no way for Flux Python commands to easily parse
TOML files without adding a new dependency or requiring Python 3.10
(which added the tomllib module to the standard library).

Vendor version 1.2.3 of the tomli library, which supposedly has the
same API as the tomllib module standard in Python 3.10. This library
can then be imported with

 try:
     import tomllib
 except ModuleNotFoundError:
     from flux.utils import tomli as tomllib

So that the vendored tomli can be removed once tomllib is virtually
everywhere.