python: add flux.util.UtilConfig class
Problem: Some Flux utilities may want to initialize and read
configuration from a hieararchy, i.e. internal defaults supplemented
by system configuration and user configuration, but no standard method
to do that exists.

Add a very simple UtilConfig class to the Python API which handles
reading configuration (initialized from internal defaults) from a
set of standard paths.

We standardize on the XDG Base Directory Specification[1] for lack
of any other standard. Even though XDG is nominally for desktop
environments, it seems other system tools are using this specification

Configuration is held as a simple dictionary, which is merged and/or
overridden in the following order:

 1. optional internal defaults
 2. /etc/xdg/flux
 3. XDG_CONFIG_DIRS
 4. $HOME/.config/flux
 5. XDG_CONFIG_HOME

As each directory is traversed, the config parser will attempt to
load all <name>.* files, where <name> is the name of the utility
reading configuration. Files with the extensions `.toml`, `.yaml`
and `.json` are supported. Other files are ignored.

A user of the UtilConfig class may override the empty validate()
method in a subclass if some validation of each loaded file is to
be performed.

Each successfully loaded file is merged with the current configuration,
overriding existing keys or adding new keys, such that the final
traversal (user configuration) has the highest precedence.

[1] https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html