flux-jobs: support named format strings from config files
Problem: It is painful to pass an entire format string to the
flux-jobs -o, --format option, especially when the format string
might be something a user or admin wants to often use.

Allow named formats to be registered to the flux-jobs command via a
set of configuration files loaded by the flux.util.UtilConfig() config
parser. This allows internal defaults to be enhanced or overridden
by system and/or user configuration. The config files follow the XDG
system base directory format so config files can be in

  - /etc/xdg/flux/flux-jobs.*
  - XDG_CONFIG_DIRS/flux-jobs.*
  - $HOME/.config/flux/flux-jobs.*
  - XDG_CONFIG_HOME/flux-jobs.*

Any of TOML, YAML, or JSON files are supported. New formats are
specified in a "formats" dictionary or table, with a named key per
format pointing to a dictionary with keys:

 - format: (required) the format string
 - description: (optional) a description of the format, displayed
    with `flux jobs --format=help`