flux-jobs: support named format strings from config files
Problem: It is painful to pass an entire format string to the
flux-jobs -o, --format option, especially when the format string
might be something a user or admin wants to often use.

Allow named formats to be registered to the flux-jobs command via
a set of configuration files loaded by the flux.util.UtilConfig()
config parser. This allows the addition of system and/or user
configuration. User configuration overrides system configuration,
but internal format names cannot be overridden.

The config files follow the XDG system base directory format so config
files can be in /etc/xdg/flux/flux-jobs.* (or under XDG_CONFIG_DIRS)
for system config, or $HOME/.config/flux/flux-jobs.* (or under
XDG_CONFIG_HOME) for user configuration.

Any of TOML, YAML, or JSON files are supported. New formats are
specified in a "formats" dictionary or table, with a named key per
format pointing to a dictionary with keys:

 - format: (required) the format string
 - description: (optional) a description of the format, displayed
    with `flux jobs --format=help`