flux-jobs: support named format strings
Problem: It would be nice to have a set of builtin flux-jobs format
strings, but this is currently not supported.

Add support for named format strings that can be provided to the
flux-jobs --format option. Allow the list of supported names to
be displayed with `flux jobs --format=help`. Register the current
internal default format string as "default."