broker: fix use-after-free segfault
Problem: on Ubuntu 22.04, t0014-runlevel.t triggers a segfault
when it runs the broker with broker.rc3_path=/bin/false.

If a module has not been fully unloaded by the time modhash_destroy()
is called, the module's context could be destroyed before its service
entry is removed.  Then a disconnect message sent when another module
context is destroyed may cause the destroyed module context to be
dereferenced.

Add code to ensure the service hash is removed when the module is
destroyed, if that hasn't happened already.

Fixes #4564.