docker: add namespaced builds for spack/ubuntu
We currently build one set of containers that are
spack focused, and ideally we should provide this
along with more vanilla (e.g., ubuntu) builds.
This update will move the spack builds into a
spack subdirectory and add two ubuntu builds -
one a base container, and the other that just
runs a make check-prep to run some tests.
This ubuntu base could be used for other contexts
of testing, and I think after the python work is
integrated we would migrate the container there
into this directory too.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>