python: add filter_empty method to JobInfoFormat class
Problem: In some cases it may be useful to have a field in the
flux-jobs format string that is only displayed when at least one job
results in a non-empty string for that field. For example, it may
make sense to add the job queue name to the default format string
for flux-jobs, but most non-system Flux instances will be using the
anonymous (unnamed) queue, so this field would always be empty.

Add a JobInfoFormat.filter_empty() method which will optionally remove
fields which format to an empty string for every job in a provided
list of JobInfo objects. The function only performs this action on
format fields that are prefixed with the indicator "?:".

This allows a format string to be crafted that will collapse selected
fields if the field does not apply to all jobs in the list, e.g.
"{id.f58} ?:{queue}".