flux-jobs: support --format=get-config=name
Problem: Users may desire a starting configuration from which to
extend or modify an existing flux-jobs format, but there is no easy
way for users to do that.

Add a special syntax for the -o, --format option to dump a
configuration snippet for a given format, e.g.

 flux jobs --format=get-config=default

will dump a TOML entry for the default format.