job-list: fix job name filter corner case
Problem: A logic error exists with job name filtering.  If a
job name is specified for filtering, filtering logic would check
if `job->name` is not-NULL and then check for string equality.
This is incorrect, as it ignores the possibility that the job->name
is not set (i.e. job->name == NULL).  If the job->name was NULL,
a job could incorrectly "pass" the filter.

Solution: Update logic to check for NULL `job->name` and reject
it in filtering if necessary.