python: fix flux.util.OutputFormat header_format()
Problem: The code that tries to generate a string-only format spec
for headers in OutputFormat.header_format() sometimes strips alignment
and width specifiers from the format spec.

Generalize the re.match() to re.search() so that Python format spec can
be found anywhere in the string, not just the beginning.

Then pass the match capture as the format spec for the header format
string.