python: JobInfoFormat: allow format spec with '!d' conversions
Problem: When converting timestamps to datetime objects with the 'd'
conversion specifier in flux-jobs output format strings, width and
alignment cannot be specified since these are not supported by the
Python datetime __format__ method.

Add a JobDatetime() subclass of datetime which overrides the __format__
method to support addition of alignment and width specifiers after a
`::` separator, e.g.:

 >>> "{0:%c::^30}".format(JobDatetime.now())
 '   Wed Sep 21 20:13:47 2022   '

Update the 'd' conversion specifier to convert timestamps to
JobDatetime instead of datetime to take advantage of JobDatetime's
support for width and alignment.