job-list: handle per-resource "cores" special case
Problem: Under several special cases to work around Jobspec V1
limitations, the shell per-resource options are used to specify
number of tasks in a job.  The per-resource type of "node" was
previously handled, but the type of "core" was not.

Solution: Handle per-resource "core" special case and set
ntasks accordingly if it is set.

Fixes #4593