flux-top: display a QUEUE column if queues are configured
Problem: flux-top does not include a QUEUE column, but this may be
useful when queues are configured in the target Flux instance.

Check if any job in the current job list defines a queue and, if so,
display a QUEUE column with the queue name. Otherwise, do not bother
displaying job queues.

Fixes #4442