flux-jobs: add queue to builtin format strings
Problem: flux-jobs does not display the job queue name in either
of the builtin formats (`default` and `deps`), but when queues are
configured this will be essential information.

Add `?:{queue}` to each of the builtin flux-jobs formats so that
a QUEUE column is displayed when queues are configured, but is not
included in the common case of a single anonymous queue.

Fixes #4602