python: add default JobFrobnicator plugins
Problem: the job frobnicator has no default plugins, but we may
want to automatically activate the frobnicator when certain config
options are present, and defaults would make that simpler.

Load 'defaults' and 'constraints' plugins if none are specified.

Add test to t2112-job-ingest-frobnicator.t to check that default
plugins are loaded.