job-ingest: enable frobnicator as needed
Problem: it is easy to configure [queues] or jobspec defaults
and forget to enable the frobnicator, which is required for
those to work.

Change the ingest pipeline logic so the frobnicator is enabled
if one of those is specified.

Add a disable flag like the validator's so that its auto-enablement
can be bypassed if needed.

Fixes #4598