job-exec: do not force nonzero exit code for lost shells
Problem: The job-exec/bulk-exec code always sets exit code 68 for
shells that are lost (i.e. node or broker failure), but this causes
all instances with a node failure to exit with "failure", which is
not always appropriate.

Set a 0 exit code for the shell on EHOSTUNREACH and let success/failure
of the job as a whole be later determined by other shell exit codes.