shell: add exception plugin and shell.lost callback
Problem: When a job shell is lost, e.g. due to node failure, the
missing job shell will never be able to send EOF to the rank 0 output
plugin.  This causes an indefinite hang on normal exit, i.e. when the
job is not immediately terminated on node failure, because the output
plugin refcount can never reach 0 and release its completion reference.

Add an exception plugin to the job shell which registers an `exception`
RPC endpoint to the rank 0 job shell. The job-exec system can then
notify the shell directly when a job shell is lost (or in the future
for other non-fatal exceptional conditions).

When the exception plugin is notified of a lost shell, it then calls
a `shell.lost` plugin callback so that other plugins can take action
on this event.

In the output plugin, add a handler for the `shell.lost` callback
which decrements the output refcount to avoid waiting forever for EOF
from a lost shell.

Fixes #4581