python: require job queue if [queues] configured
Problem: if queues are configured, but no default queue is configured,
and a job is submitted without a queue, the job is ingested and
allowed to fail later

Modify the 'defaults' frobnicator plugin to reject such a job at ingest.

Fix sharness limit tests that were using queue configurations that are now
considered invalid.

Fixes #4614