flux-start: add --test-pmi-clique=per-broker
Problem: It would be useful to start a test instance with flux-start
using a broker mapping simulating multiple nodes, but flux-start does
not support this, the mapping always places all brokers on one node.

Support a "per-broker" argument to the flux-start --test-pmi-clique
option, which will cause flux-start to create a mapping with a clique
per broker, i.e. a separate node for each broker.