broker: add general parental notification function
Problem: The broker may need to contact its parent (when one exists)
for other purposes than just sending a job memo with the instance
URI, but the handle to the parent is opened and closed within the
set_uri_job_memo() function, which doesn't lend itself to easily
adding other parent interactions.

Add a parental notification function which calls the set_uri_job_memo()
function, but allows for other RPCs to be initiated while the parent
handle is open.