broker: overlay: add overlay_get_default_critical_ranks
Problem: The broker will need to obtain the set of so-called "critical"
ranks from the overlay topology, but no such function exists for
this purpose.

Add overlay_get_default_critical_ranks() to return an idset of ranks
which the overlay by default deems "critical" for instance operation.

For now, return rank 0 plus the set of internal ranks from the
topology, since these ranks are used in routing messages.