broker: notify parent of critical nodes
Problem: The job-exec module supports updating the set of ranks
which will generate a fatal job exception when they are lost, but the
broker does not currently support this. Therefore, jobs are killed
when any node in the job is lost, even though jobs that are also
instances of Flux could easily continue running after the loss of
non-critical nodes.

Calculate the set of critical *shell* ranks (or: nodeids) using the
broker.critical-ranks and broker.mapping attributes.  Notify the
job-exec service in the parent instance of this possibly reduced
set of critical ranks via the job-exec.critical-ranks RPC.
Do this in parallel with the existing job memo uri update in
execute_parental_notifications().

This will allow all jobs that are also Flux instances to tolerate
the loss of leaf nodes and continue running jobs if possible.

Fixes #4589